+++ Waiter/SoupWaiterAdmin.php
@@ -395,7 +395,7 @@ class SoupWaiterAdmin extends SitePersistedSingleton {
 		$error_obj = false;
 
 		$newpost = $_POST;
-		$newpost['post_content'] .= "<p>Created by ".SoupWaiter::single()->owner_name."</p>";
+		$newpost['post_content'] .= "<p class='autocreated'>Created by ".SoupWaiter::single()->owner_name."</p>";
 
 		$postId = wp_insert_post($newpost,$error_obj);
 		if (!$error_obj){
@@ -403,6 +403,9 @@ class SoupWaiterAdmin extends SitePersistedSingleton {
 			wp_set_post_categories( $postId, $_POST['cats']);
 			set_post_thumbnail($postId,$_POST['featured_image']);
 			update_post_meta($postId,'topic',$_POST['topic']);
+			update_post_meta($postId,'latitude',$_POST['latitude']);
+			update_post_meta($postId,'longitude',$_POST['longitude']);
+			update_post_meta($postId,'destination_id',$_POST['destination_id']);
 		}
 		SoupWaiter::single()->wp_async_save_post($postId,get_post($postId));
 		// $this->persist(); // Don't know why we were persisting SoupWaiterAdmin, not needed, not here anyway
@@ -438,6 +441,8 @@ class SoupWaiterAdmin extends SitePersistedSingleton {
         // In case the destination is multiple words
 
 		$context['permTags'][] = 'VacationSoup';
+		// Allows rendering the lat/long
+		$context['destination'] = SoupWaiter::single()->get_destination();
         foreach (SoupWaiter::single()->destinations_for_property as $destination) {
 	        $render = '';
 	        foreach (explode(' ',preg_replace('/[^\w]+/ui', ' ', $destination['rendered'])) as $word){
@@ -526,30 +531,54 @@ class SoupWaiterAdmin extends SitePersistedSingleton {
      */
 	public function get_properties(){
         $propertyCount = SoupWaiter::single()->property_count;
-        $properties = [];
+        static $properties = [];
         $destinations = []; // Might as well set these up too
+		$deduped = [];
         $dCount = 0;
 
-        for ($i=0;$i<($propertyCount);$i++){
-            $property = Property::findOne($i);
-            $properties[] = $property;
+        if (empty ($properties)) {
+	        for ( $i = 0; $i < ( $propertyCount ); $i ++ ) {
+		        $property     = Property::findOne( $i );
+		        $properties[] = $property;
 
-	        if (!empty($property->destination)){
-		        $destinations[$dCount] = ['id'=>$dCount,'property'=>$i,'rendered'=>"{$property->join} {$property->destination}",'destination'=>"{$property->destination}"];
-		        $dCount++;
-	        }
-	        if (!empty($property->destination2)){
-		        $destinations[$dCount] = ['id'=>$dCount,'property'=>$i,'rendered'=>"{$property->join2} {$property->destination2}",'destination'=>"{$property->destination2}"];
-		        $dCount++;
-	        }
-	        if (!empty($property->destination3)){
-		        $destinations[$dCount] = ['id'=>$dCount,'property'=>$i,'rendered'=>"{$property->join3} {$property->destination3}",'destination'=>"{$property->destination3}"];
-		        $dCount++;
-	        }
+		        if ( ! empty( $property->destination ) && !isset($deduped[$property->destination])) {
+			        $deduped[$property->destination] = $destinations[ $dCount ] = [
+				        'id'          => $dCount,
+				        'property'    => $i,
+				        'latitude'    => "{$property->latitude}",
+				        'longitude'   => "{$property->longitude}",
+				        'rendered'    => "{$property->join} {$property->destination}",
+				        'destination' => "{$property->destination}"
+			        ];
+			        $dCount ++;
+		        }
+		        if ( ! empty( $property->destination2 )  && !isset($deduped[$property->destination2])) {
+			        $deduped[$property->destination] = $destinations[ $dCount ] = [
+				        'id'          => $dCount,
+				        'property'    => $i,
+				        'latitude'    => "{$property->latitude}",
+				        'longitude'   => "{$property->longitude}",
+				        'rendered'    => "{$property->join2} {$property->destination2}",
+				        'destination' => "{$property->destination2}"
+			        ];
+			        $dCount ++;
+		        }
+		        if ( ! empty( $property->destination3 )  && !isset($deduped[$property->destination3])) {
+			        $deduped[$property->destination] = $destinations[ $dCount ] = [
+				        'id'          => $dCount,
+				        'property'    => $i,
+				        'latitude'    => "{$property->latitude}",
+				        'longitude'   => "{$property->longitude}",
+				        'rendered'    => "{$property->join3} {$property->destination3}",
+				        'destination' => "{$property->destination3}"
+			        ];
+			        $dCount ++;
+		        }
 
+	        }
+	        sort($destinations);
+	        SoupWaiter::single()->destinations = $destinations;
         }
-        sort(array_unique($destinations));
-        SoupWaiter::single()->destinations = $destinations;
         return $properties;
     }
 	/**