+++ soup-waiter.php
@@ -47,6 +47,13 @@ function soup_waiter_register_required_plugins() {
 			'required'  => true,
 			'version'            => '1.4.1', // E.g. 1.0.0. If set, the active plugin must be this version or higher. If the plugin version is higher than the plugin version installed, the user will be notified to update the plugin.
 			'force_activation'   => true, // If true, plugin is activated upon theme activation and cannot be deactivated until theme switch.
+		],
+		[
+			'name'      => 'Companion Auto Update',
+			'slug'      => 'companion-auto-update',
+			'required'  => false,
+			'version'            => '2.9.6', // E.g. 1.0.0. If set, the active plugin must be this version or higher. If the plugin version is higher than the plugin version installed, the user will be notified to update the plugin.
+			'force_activation'   => false, // If true, plugin is activated upon theme activation and cannot be deactivated until theme switch.
 		]
 	];
 	$config = [